Keep all workers and send them new data

+++ app.js
@@ -20,42 +20,34 @@ const ffmpeg = require('fluent-ffmpeg');
 ffmpeg.setFfmpegPath(ffmpegPath);
 
 const CPUS = os.cpus();
-const workers = [];
 const results = [];
 let images = [];
 
-const workerEvent = new EventEmitter();
+const getNextImage = () => {
+    return images.shift()
+}
 
 const createWorker = () => {
   return new Promise(resolve => {
-    const image = images.shift();
+    const image = getNextImage()
 
     if (!image) {
       resolve();
     }
 
     const worker = childProcess.fork(path.resolve(__dirname, 'worker.js'), ['--image', image]);
+    
     worker.on('message', ({ clusters }) => {
       results.push(clusters);
-    });
+      const nextImage = getNextImage()
 
-    worker.on('exit', () => {
-      const workerIndex = workers.indexOf(worker);
-      worker.kill();
-
-      if (workerIndex > -1) workers.splice(workerIndex, 1);
-
-      if (images.length === 0) {
-        if (workers.length === 0) {
-          workerEvent.emit('handle:quantification');
-        }
-        resolve();
+      if(nextImage) {
+          worker.send({image: nextImage})
+      } else {
+          worker.kill()
+          resolve()
       }
-
-      return createWorker();
     });
-
-    workers.push(worker);
   });
 };
 
@@ -90,15 +82,11 @@ const processScreenshots = () => {
       });
 
       Promise.all(promises).then(() => {
-        console.log('Promise finished');
+          handleQuantification()
       });
     });
   });
 };
-
-workerEvent.once('handle:quantification', () => {
-  handleQuantification();
-});
 
 console.time('Timer');
 getScreenshotsFromVideo(args.video)
